@@ -4,11 +4,11 @@ import cz.tefek.botdiril.userdata.icon.Icons;
 
 public enum EnumCurrency
 {
-    XP("xp", Icons.ACHIEVEMENT_BOT, "Experience", "The main cz.tefek.botdiril.command.currency of Botdiril! Collect experience to unlock more items, loot and commands."),
+    XP("xp", Icons.ACHIEVEMENT_BOT, "Experience", "The main currency of Botdiril! Collect experience to unlock more items, loot and commands."),
     COINS("coin", Icons.COIN, "Coin", "Used mostly to buy goods and exchange for other currencies."),
-    KEKS("kek", Icons.KEK, "Kek", "The main cz.tefek.botdiril.command.gambling cz.tefek.botdiril.command.currency, can be exchanged for Kek Tokens."),
-    TOKENS("kektoken", Icons.TOKEN, "Kek Token", "Allow you to get some sweet loot for your cz.tefek.botdiril.command.gambling activities."),
-    MEGAKEKS("megakek", Icons.MEGAKEK, "MegaKek", "Keks but on steroids. Reach one googol MegaKeks in no time!"),
+    KEKS("kek", Icons.KEK, "Kek", "The main gambling currency, can be exchanged for Kek Tokens."),
+    TOKENS("kektoken", Icons.TOKEN, "Kek Token", "Allow you to get some sweet loot for your gambling activities."),
+    MEGAKEKS("megakek", Icons.MEGAKEK, "MegaKek", "Keks but on steroids. Handle with care."),
     DUST("dust", Icons.DUST, "Dust", "Used mostly for crafting, Dust allows you to recycle duplicate items for new goods."),
     KEYS("key", Icons.KEY, "Key", "Unlock even more loot paths.");
 
